city and district detail added for delete and create/edit

+++ NHRMS_WebAPI/Models/TechnicalErrorStatus.cs
@@ -17,7 +17,7 @@ namespace NHRMS_WebAPI.Models
         public string Remarks { get; set; }
         public string ApprovalDateTime { get; set; }
         public string EmployeeDesignation { get; set; }
-        public int EmpBranchID { get; set; }
+        public long EmpBranchID { get; set; }
         public string EmpBranchName { get; set; }
         public string EmpOfficeName { get; set; }
         public string AttachmentDocument { get; set; }
